daemon: move errors-related code from response.go into errors.go
